refactor(workflows): update imports and types after restructure
update workflow runtime and tests to use new import paths for
sessions/tasks modules and WorkflowState type name
