+++ typescript/lib/agent-node/tests/workflow-runtime.int.test.ts
@@ -9,7 +9,7 @@ import { describe, it, expect, afterEach } from 'vitest';
 
 import { WorkflowPluginLoader } from '../src/config/runtime/workflow-loader.js';
 import type { EffectiveWorkflow } from '../src/config/composers/effective-set-composer.js';
-import type { WorkflowYield } from '../src/workflows/types.js';
+import type { WorkflowState } from '../src/workflows/types.js';
 import { createTestConfigWorkspace } from './utils/test-config-workspace.js';
 
 describe('Workflow Runtime Integration Tests', () => {
@@ -294,7 +294,7 @@ export default {
         parameters: {},
       };
 
-      const yields: WorkflowYield[] = [];
+      const yields: WorkflowState[] = [];
       for await (const result of loadedPlugin!.plugin.execute(context)) {
         yields.push(result);
       }
@@ -378,7 +378,7 @@ export default {
         taskId: 'test-task',
       };
 
-      const yields: WorkflowYield[] = [];
+      const yields: WorkflowState[] = [];
       for await (const result of loadedPlugin!.plugin.execute(context)) {
         yields.push(result);
       }
@@ -457,7 +457,7 @@ export default {
         },
       };
 
-      const yields: WorkflowYield[] = [];
+      const yields: WorkflowState[] = [];
       for await (const result of loadedPlugin!.plugin.execute(context)) {
         yields.push(result);
       }
@@ -534,7 +534,7 @@ export default {
         parameters: loadedPlugin!.overrides, // Apply overrides
       };
 
-      const yields: WorkflowYield[] = [];
+      const yields: WorkflowState[] = [];
       for await (const result of loadedPlugin!.plugin.execute(context)) {
         yields.push(result);
       }
